Update per review comments, add to unit test
- moved getIconResumed() from SdlProxyALM to SdlProxyBase
- add setIconResumed() to RegisterAppInterfaceResponse
- add corresponding test in RegisterAppInterfaceResponse unit test
