+++ sdl_android/src/androidTest/assets/json/RegisterAppInterface.json
@@ -192,7 +192,8 @@
         0,
         1,
         2
-      ]
+      ],
+      "iconResumed":true
     }
   }
 }
